refactor(handler): restructure http handler layer
