+++ internal/adapter/handler/http/helper.go
@@ -1,4 +1,4 @@
-package handler
+package http
 
 import (
 	"strconv"
