+++ internal/adapter/handler/http/middleware.go
@@ -1,4 +1,4 @@
-package handler
+package http
 
 import (
 	"strings"
