+++ internal/adapter/handler/http/response.go
@@ -1,4 +1,4 @@
-package handler
+package http
 
 import (
 	"net/http"
