+++ internal/adapter/handler/http/router.go
@@ -1,4 +1,4 @@
-package handler
+package http
 
 import (
 	"fmt"
